add LB to TMC1A

+++ reduction_utils3.py
@@ -1307,10 +1307,14 @@ def self_calibrate(prefix,data_params,mode='SB-only',iteration=0,selfcalmode='p'
                   SB_spwmap=[0,0,0,0,0,0,0],LB_contspws='',LB_spwmap=[0,0,0,0,0,0,0],
                   cellsize=None,imsize=None,scales=None,finalimageonly=False,remove_all_following_iterations=True,         
                   sidelobethreshold=2.5,noisethreshold=5.0,lownoisethreshold=1.5,smoothfactor=1.0,combine='spw',skipImage=False):
+   ### Use skipImage with care
    contspws=''
    spwmap=[0,0,0,0,0]
    if remove_all_following_iterations:   ### Remove all following selfcal files after the iteration being run
-      os.system("rm -rf "+"*_"+mode+"*p{{{0:d}..99}}*".format(iteration))
+      if skipImage:   # don't erase the current iteration if you want to skip imaging and re-run with current model
+         os.system("rm -rf "+"*_"+mode+"*p{{{0:d}..99}}*".format(iteration+1))
+      else:
+         os.system("rm -rf "+"*_"+mode+"*p{{{0:d}..99}}*".format(iteration))
    if prevselfcalmode==None:
       prevselfcalmode=selfcalmode
 
@@ -1337,27 +1341,30 @@ def self_calibrate(prefix,data_params,mode='SB-only',iteration=0,selfcalmode='p'
       if (mode =='LB-only') and ('LB' in i): # skip over SB EBs if in LB-only mode
          continue
       if (mode =='LB+SB') and ('SB' in i) and (iteration == 0): # slightly different flow for the SB data in iter0
-         os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
-         os.system('cp -r '+data_params[i]['vis_avg_selfcal']+' '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
-         data_params[i]['selfcal_spwmap']=data_params[i]['selfcal_spwmap_SB-only']
-         data_params[i]['selfcal_tables']=data_params[i]['selfcal_tables_SB-only']
+         if not skipImage:
+            os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
+            os.system('cp -r '+data_params[i]['vis_avg_selfcal']+' '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
+            data_params[i]['selfcal_spwmap']=data_params[i]['selfcal_spwmap_SB-only']
+            data_params[i]['selfcal_tables']=data_params[i]['selfcal_tables_SB-only']
          vislist.append(data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
          continue
       if iteration == 0:
-         os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
-         os.system('cp -r '+data_params[i]['vis_avg_shift_rescaled']+' '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
-         data_params[i]['selfcal_tables']=[]
-         data_params[i]['selfcal_spwmap']=[]
+         if not skipImage:
+            os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
+            os.system('cp -r '+data_params[i]['vis_avg_shift_rescaled']+' '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_p0.ms'))
+            data_params[i]['selfcal_tables']=[]
+            data_params[i]['selfcal_spwmap']=[]
       elif iteration > 0:
          if len(data_params[i]['selfcal_tables']) > iteration:  ### remove selfcal table entries when redoing self-cal stages
             data_params[i]['selfcal_tables'] = data_params[i]['selfcal_tables'][0:iteration]
             data_params[i]['selfcal_spwmap'] = data_params[i]['selfcal_spwmap'][0:iteration]
-         print('Input MS: ',data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+prevselfcalmode+str(iteration-1)+'.ms'))
-         os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
-         split(vis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+prevselfcalmode+str(iteration-1)+'.ms'),
-               outputvis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
-         data_params[i]['vis_avg_selfcal']=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms')
-         print('Output MS: ',data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
+         if not skipImage:
+            print('Input MS: ',data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+prevselfcalmode+str(iteration-1)+'.ms'))
+            os.system('rm -rf '+data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
+            split(vis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+prevselfcalmode+str(iteration-1)+'.ms'),
+                  outputvis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
+            data_params[i]['vis_avg_selfcal']=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms')
+            print('Output MS: ',data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
       vislist.append(data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'))
 
 
@@ -1394,7 +1401,9 @@ def self_calibrate(prefix,data_params,mode='SB-only',iteration=0,selfcalmode='p'
          gaincal(vis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.ms'), 
                     caltable=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.g'), 
                     gaintype='T', spw=contspws,refant=data_params[i]["refant"], calmode=selfcalmode, solint=solint, 
-                    minsnr=2.0, minblperant=4,combine=combine,solnorm=True)
+                    minsnr=4.0, minblperant=4,combine=combine,solnorm=True)
+         flagdata(vis=data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.g'),mode='clip',
+                  clipoutside=True,clipminmax =[0.5,1.5],datacolumn='CPARAM')
       # Add caltable to the dictionary for use later
       data_params[i]['selfcal_tables'].append(data_params[i]['vis_avg_shift_rescaled'].replace('.ms','_'+mode+'_'+selfcalmode+str(iteration)+'.g'))
